📝🩹 fix the RtD build (#105)
## Description

The RtD build has been broken for a while because the PDF build (which
is not tested on PRs) would fail. This PR simply disables the PDF
builds.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.